requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Werkzeug>=3.1.0
tensorflow==2.17.0
Pillow>=10.0
numpy>=1.23.5,<2.0.0
opencv-python-headless==4.11.0.86
python-dotenv